+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed null deref caused by Echo Chamber trying to spawn invalid projectiles due to an invalid base projectile or invalid player owner
 - Fixed null deref caused by Display Pedestal trying to update the status of a gun that no longer exists
 - Fixed null deref caused by Paintball Cannon trying to apply its affects to certain non-Paintball Cannon projectiles
+- Added some temporary patches to fix a handful of compatibility issues with JuneLib until it gets properly updated
 
 ## 1.28.1 (2025-10-01)